requirements.txt
requests~=2.31.0
pyserial~=3.5

# 阿里云语音识别需要的
oss2~=2.18.6
aliyun-python-sdk-core>=2.13.3
dashscope~=1.20.1
pyinstaller~=6.9.0
setuptools~=68.2.0
PyYAML~=6.0.1